refactor circe encoders
